requirements.txt
flask
google-generativeai
requests
python-dotenv